feat: Add debug map display during gameplay
The commit adds a debug feature to display the map of the computer during gameplay. This feature is controlled by the `DEBUG` flag. If `DEBUG` is set to true, the map will be displayed along with other game information. If `DEBUG` is false, the map will not be displayed. This change improves debugging and testing capabilities.